Getting started with cilk. Race bugs detected
At first I tried to use only the cilk for statement and see what is going to happen. Apparently the execution time was slower than the sequential, because I didn't set any number of workers to my system. After doing that, race bugs detected using the appropriate flag. So we need to resolve them. In general cilk plus is deprecated but it has good documentation, you can refer to Intel's reference manual